// gen/ts/models/GPS.ts
/* tslint:disable */
/* eslint-disable */
/**
 * StarRoll API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * GPS location
 * @export
 * @interface GPS
 */
export interface GPS {
    /**
     * 
     * @type {number}
     * @memberof GPS
     */
    longitude: number;
    /**
     * 
     * @type {number}
     * @memberof GPS
     */
    latitude: number;
}

/**
 * Check if a given object implements the GPS interface.
 */
export function instanceOfGPS(value: object): value is GPS {
    if (!('longitude' in value) || value['longitude'] === undefined) return false;
    if (!('latitude' in value) || value['latitude'] === undefined) return false;
    return true;
}

export function GPSFromJSON(json: any): GPS {
    return GPSFromJSONTyped(json, false);
}

export function GPSFromJSONTyped(json: any, ignoreDiscriminator: boolean): GPS {
    if (json == null) {
        return json;
    }
    return {
        
        'longitude': json['longitude'],
        'latitude': json['latitude'],
    };
}

export function GPSToJSON(json: any): GPS {
    return GPSToJSONTyped(json, false);
}

export function GPSToJSONTyped(value?: GPS | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'longitude': value['longitude'],
        'latitude': value['latitude'],
    };
}